--- src/models/ammo.rs
@@ -1,7 +1,0 @@
-pub struct Ammo {
-    pub caliber: String,
-    pub damage: u64,
-    pub armor_damage: u64,
-    pub penetration_power: u64,
-    pub recoil: u64,
-}
