--- src/models/mod.rs
@@ -1,3 +1,0 @@
-pub mod ammo;
-pub mod quest;
-pub mod task;
