--- src/models/quest.rs
@@ -1,7 +1,0 @@
-use serde::Deserialize;
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct Quest {
-    pub id: String,
-    pub name: String,
-}
