--- src/utils/mod.rs
@@ -1,3 +1,0 @@
-pub mod graphql;
-pub mod loader;
-pub mod response;
